make fixtures more realistic shape
