+++ packages/go/test/fixtures/29-gin-server-go1.26/cmd/server/main.go
@@ -1,0 +1,9 @@
+package main
+
+import "gin-server-go126/internal/server"
+
+func main() {
+	if err := server.Run(); err != nil {
+		panic(err)
+	}
+}
