+++ packages/go/test/fixtures/29-gin-server-go1.26/internal/server/server.go
@@ -1,15 +1,14 @@
-package main
+package server
 
 import (
 	"embed"
-	"encoding/json"
 	"io/fs"
 	"net/http"
 	"os"
 	"runtime"
 	"time"
 
-	"github.com/go-chi/chi/v5"
+	"github.com/gin-gonic/gin"
 )
 
 //go:embed assets/*
@@ -36,51 +35,57 @@ type ItemResponse struct {
 	Timestamp string   `json:"timestamp"`
 }
 
-func main() {
+func Run() error {
+	gin.SetMode(gin.ReleaseMode)
+	r := gin.New()
+	r.Use(gin.Recovery())
+
 	assetFS, err := fs.Sub(assets, "assets")
 	if err != nil {
-		panic(err)
+		return err
 	}
 
-	r := chi.NewRouter()
-	r.Handle("/static/*", http.StripPrefix("/static/", http.FileServer(http.FS(assetFS))))
+	r.StaticFS("/static", http.FS(assetFS))
 
-	r.Get("/", func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		_, _ = w.Write([]byte("<html><body>framework:chi version:" + runtime.Version() + " randomness:RANDOMNESS_PLACEHOLDER</body></html>"))
+	r.GET("/", func(c *gin.Context) {
+		c.Data(
+			http.StatusOK,
+			"text/html; charset=utf-8",
+			[]byte("<html><body>framework:gin version:"+runtime.Version()+" randomness:RANDOMNESS_PLACEHOLDER</body></html>"),
+		)
 	})
 
-	r.Get("/api/data", func(w http.ResponseWriter, r *http.Request) {
+	r.GET("/api/data", func(c *gin.Context) {
 		items := []DataItem{
 			{ID: 1, Name: "Sample Item 1", Value: 100},
 			{ID: 2, Name: "Sample Item 2", Value: 200},
 			{ID: 3, Name: "Sample Item 3", Value: 300},
 		}
 
-		writeJSON(w, http.StatusOK, DataResponse{
+		c.JSON(http.StatusOK, DataResponse{
 			Data:      items,
 			Total:     len(items),
-			Framework: "chi",
+			Framework: "gin",
 			Version:   runtime.Version(),
 			Timestamp: time.Now().UTC().Format(time.RFC3339),
 		})
 	})
 
-	r.Get("/api/items/{id}", func(w http.ResponseWriter, r *http.Request) {
-		writeJSON(w, http.StatusOK, ItemResponse{
+	r.GET("/api/items/:id", func(c *gin.Context) {
+		c.JSON(http.StatusOK, ItemResponse{
 			Item: DataItem{
 				ID:    1,
-				Name:  "Sample Item " + chi.URLParam(r, "id"),
+				Name:  "Sample Item " + c.Param("id"),
 				Value: 100,
 			},
-			Framework: "chi",
+			Framework: "gin",
 			Version:   runtime.Version(),
 			Timestamp: time.Now().UTC().Format(time.RFC3339),
 		})
 	})
 
-	r.NotFound(func(w http.ResponseWriter, r *http.Request) {
-		http.Error(w, "not found:chi", http.StatusNotFound)
+	r.NoRoute(func(c *gin.Context) {
+		c.String(http.StatusNotFound, "not found:gin")
 	})
 
 	port := os.Getenv("PORT")
@@ -88,11 +93,5 @@ func main() {
 		port = "3000"
 	}
 
-	_ = http.ListenAndServe(":"+port, r)
-}
-
-func writeJSON(w http.ResponseWriter, status int, payload any) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	_ = json.NewEncoder(w).Encode(payload)
+	return r.Run(":" + port)
 }
