+++ packages/go/test/fixtures/31-chi-server-go1.26/cmd/api/main.go
@@ -1,0 +1,9 @@
+package main
+
+import "chi-server-go126/internal/api"
+
+func main() {
+	if err := api.Run(); err != nil {
+		panic(err)
+	}
+}
