+++ packages/go/test/fixtures/31-chi-server-go1.26/internal/api/assets/info.txt
